+++ modules/generate-mongo-id/controller.js
