+++ modules/generate-mongo-id/init.js
