+++ modules/generate-mongo-id/routes.js
